# ARROW-11300: [Rust][DataFusion] Further performance improvements on hash aggregation with small groups