--- a/rust/datafusion/src/physical_plan/hash_aggregate.rs
+++ b/rust/datafusion/src/physical_plan/hash_aggregate.rs
@@ -358,6 +358,7 @@ fn group_aggregate_batch(
                     accumulator.merge_batch(&values)
                 }
             })
+            // 2.5
             .and({
                 indices.clear();
                 Ok(())