--- a/rust/datafusion/src/physical_plan/hash_aggregate.rs
+++ b/rust/datafusion/src/physical_plan/hash_aggregate.rs
@@ -30,28 +30,27 @@ use crate::error::{DataFusionError, Result};
 use crate::physical_plan::{Accumulator, AggregateExpr};
 use crate::physical_plan::{Distribution, ExecutionPlan, Partitioning, PhysicalExpr};
 
+use super::{ expressions::Column, group_scalar::GroupByScalar,
+    RecordBatchStream, SendableRecordBatchStream,
+};
+use ahash::RandomState;
+use arrow::array::{Array, UInt32Builder};
+use arrow::datatypes::{DataType, Field, Schema, SchemaRef, TimeUnit};
 use arrow::error::{ArrowError, Result as ArrowResult};
 use arrow::record_batch::RecordBatch;
 use arrow::{
     array::BooleanArray,
-    datatypes::{DataType, Field, Schema, SchemaRef, TimeUnit},
 };
 use arrow::{
     array::{
-        ArrayRef, Float32Array, Float64Array, Int16Array, Int32Array, Int64Array,
-        Int8Array, StringArray, UInt16Array, UInt32Array, UInt64Array, UInt8Array,
+        ArrayRef, Float32Array, Float64Array, Int16Array, Int32Array, Int64Array, Int8Array,
+        StringArray, UInt16Array, UInt32Array, UInt64Array, UInt8Array,
     },
     compute,
 };
-use pin_project_lite::pin_project;
-
-use super::{
-    expressions::Column, group_scalar::GroupByScalar, RecordBatchStream,
-    SendableRecordBatchStream,
-};
-use ahash::RandomState;
 use hashbrown::HashMap;
 use ordered_float::OrderedFloat;
+use pin_project_lite::pin_project;
 
 use arrow::array::{TimestampMicrosecondArray, TimestampNanosecondArray};
 use async_trait::async_trait;
@@ -322,48 +321,73 @@ fn group_aggregate_batch(
             });
     }
 
+    // Collect all indices + offsets based on keys in this vec
+    let mut batch_indices: UInt32Builder = UInt32Builder::new(0);
+    let mut offsets = vec![0];
+    let mut offset_so_far = 0;
+    for key in batch_keys.iter() {
+        let (_, _, indices) = accumulators.get_mut(key).unwrap();
+        batch_indices.append_slice(&indices)?;
+        offset_so_far += indices.len();
+        offsets.push(offset_so_far);
+    }
+    let batch_indices = batch_indices.finish();
+    
+    // `Take` all values based on indices into Arrays
+    let values: Vec<Vec<Arc<dyn Array>>> = aggr_input_values
+        .iter()
+        .map(|array| {
+            array
+                .iter()
+                .map(|array| {
+                    compute::take(
+                        array.as_ref(),
+                        &batch_indices,
+                        None, // None: no index check
+                    )
+                    .unwrap()
+                })
+                .collect()
+            // 2.3
+        })
+        .collect();
+
     // 2.1 for each key in this batch
     // 2.2 for each aggregation
-    // 2.3 `take` from each of its arrays the keys' values
+    // 2.3 `slice` from each of its arrays the keys' values
     // 2.4 update / merge the accumulator with the values
     // 2.5 clear indices
-    batch_keys.iter_mut().try_for_each(|key| {
-        let (_, accumulator_set, indices) = accumulators.get_mut(key).unwrap();
-        let primitive_indices = UInt32Array::from(indices.clone());
-        // 2.2
-        accumulator_set
-            .iter_mut()
-            .zip(&aggr_input_values)
-            .map(|(accumulator, aggr_array)| {
-                (
-                    accumulator,
-                    aggr_array
-                        .iter()
-                        .map(|array| {
+    batch_keys
+        .iter_mut()
+        .zip(offsets.windows(2))
+        .try_for_each(|(key, offsets)| {
+            let (_, accumulator_set, indices) = accumulators.get_mut(key).unwrap();
+            // 2.2
+            accumulator_set
+                .iter_mut()
+                .zip(values.iter())
+                .map(|(accumulator, aggr_array)| {
+                    (
+                        accumulator,
+                        aggr_array.iter().map(|array| {
                             // 2.3
-                            compute::take(
-                                array.as_ref(),
-                                &primitive_indices,
-                                None, // None: no index check
-                            )
-                            .unwrap()
-                        })
-                        .collect::<Vec<ArrayRef>>(),
-                )
-            })
-            .try_for_each(|(accumulator, values)| match mode {
-                AggregateMode::Partial => accumulator.update_batch(&values),
-                AggregateMode::Final => {
-                    // note: the aggregation here is over states, not values, thus the merge
-                    accumulator.merge_batch(&values)
-                }
-            })
-            // 2.5
-            .and({
-                indices.clear();
-                Ok(())
-            })
-    })?;
+                            array.slice(offsets[0], offsets[1] - offsets[0])
+                        }).collect(),
+                    )
+                })
+                .try_for_each(|(accumulator, values)| match mode {
+                    AggregateMode::Partial => accumulator.update_batch(&values),
+                    AggregateMode::Final => {
+                        // note: the aggregation here is over states, not values, thus the merge
+                        accumulator.merge_batch(&values)
+                    }
+                })
+                // 2.5
+                .and({
+                    indices.clear();
+                    Ok(())
+                })
+        })?;
     Ok(accumulators)
 }
 
@@ -502,14 +526,9 @@ impl GroupedHashAggregateStream {
 
         let schema_clone = schema.clone();
         tokio::spawn(async move {
-            let result = compute_grouped_hash_aggregate(
-                mode,
-                schema_clone,
-                group_expr,
-                aggr_expr,
-                input,
-            )
-            .await;
+            let result =
+                compute_grouped_hash_aggregate(mode, schema_clone, group_expr, aggr_expr, input)
+                    .await;
             tx.send(result)
         });
 
@@ -522,16 +541,12 @@ impl GroupedHashAggregateStream {
 }
 
 type AccumulatorSet = Vec<Box<dyn Accumulator>>;
-type Accumulators =
-    HashMap<Vec<u8>, (Box<[GroupByScalar]>, AccumulatorSet, Vec<u32>), RandomState>;
+type Accumulators = HashMap<Vec<u8>, (Box<[GroupByScalar]>, AccumulatorSet, Vec<u32>), RandomState>;
 
 impl Stream for GroupedHashAggregateStream {
     type Item = ArrowResult<RecordBatch>;
 
-    fn poll_next(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Option<Self::Item>> {
+    fn poll_next(self: std::pin::Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         if self.finished {
             return Poll::Ready(None);
         }
@@ -563,10 +578,7 @@ impl RecordBatchStream for GroupedHashAggregateStream {
 }
 
 /// Evaluates expressions against a record batch.
-fn evaluate(
-    expr: &Vec<Arc<dyn PhysicalExpr>>,
-    batch: &RecordBatch,
-) -> Result<Vec<ArrayRef>> {
+fn evaluate(expr: &Vec<Arc<dyn PhysicalExpr>>, batch: &RecordBatch) -> Result<Vec<ArrayRef>> {
     expr.iter()
         .map(|expr| expr.evaluate(&batch))
         .map(|r| r.map(|v| v.into_array(batch.num_rows())))
@@ -584,9 +596,7 @@ fn evaluate_many(
 }
 
 /// uses `state_fields` to build a vec of expressions required to merge the AggregateExpr' accumulator's state.
-fn merge_expressions(
-    expr: &Arc<dyn AggregateExpr>,
-) -> Result<Vec<Arc<dyn PhysicalExpr>>> {
+fn merge_expressions(expr: &Arc<dyn AggregateExpr>) -> Result<Vec<Arc<dyn PhysicalExpr>>> {
     Ok(expr
         .state_fields()?
         .iter()
@@ -606,9 +616,7 @@ fn aggregate_expressions(
     mode: &AggregateMode,
 ) -> Result<Vec<Vec<Arc<dyn PhysicalExpr>>>> {
     match mode {
-        AggregateMode::Partial => {
-            Ok(aggr_expr.iter().map(|agg| agg.expressions()).collect())
-        }
+        AggregateMode::Partial => Ok(aggr_expr.iter().map(|agg| agg.expressions()).collect()),
         // in this mode, we build the merge expressions of the aggregation
         AggregateMode::Final => Ok(aggr_expr
             .iter()
@@ -632,8 +640,8 @@ async fn compute_hash_aggregate(
     aggr_expr: Vec<Arc<dyn AggregateExpr>>,
     mut input: SendableRecordBatchStream,
 ) -> ArrowResult<RecordBatch> {
-    let mut accumulators = create_accumulators(&aggr_expr)
-        .map_err(DataFusionError::into_arrow_external_error)?;
+    let mut accumulators =
+        create_accumulators(&aggr_expr).map_err(DataFusionError::into_arrow_external_error)?;
 
     let expressions = aggregate_expressions(&aggr_expr, &mode)
         .map_err(DataFusionError::into_arrow_external_error)?;
@@ -666,8 +674,7 @@ impl HashAggregateStream {
 
         let schema_clone = schema.clone();
         tokio::spawn(async move {
-            let result =
-                compute_hash_aggregate(mode, schema_clone, aggr_expr, input).await;
+            let result = compute_hash_aggregate(mode, schema_clone, aggr_expr, input).await;
             tx.send(result)
         });
 
@@ -718,10 +725,7 @@ fn aggregate_batch(
 impl Stream for HashAggregateStream {
     type Item = ArrowResult<RecordBatch>;
 
-    fn poll_next(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Option<Self::Item>> {
+    fn poll_next(self: std::pin::Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         if self.finished {
             return Poll::Ready(None);
         }
@@ -786,16 +790,12 @@ fn create_batch_from_map(
             let mut groups = (0..num_group_expr)
                 .map(|i| match &group_by_values[i] {
                     GroupByScalar::Float32(n) => {
-                        Arc::new(Float32Array::from(vec![(*n).into()] as Vec<f32>))
-                            as ArrayRef
+                        Arc::new(Float32Array::from(vec![(*n).into()] as Vec<f32>)) as ArrayRef
                     }
                     GroupByScalar::Float64(n) => {
-                        Arc::new(Float64Array::from(vec![(*n).into()] as Vec<f64>))
-                            as ArrayRef
+                        Arc::new(Float64Array::from(vec![(*n).into()] as Vec<f64>)) as ArrayRef
                     }
-                    GroupByScalar::Int8(n) => {
-                        Arc::new(Int8Array::from(vec![*n])) as ArrayRef
-                    }
+                    GroupByScalar::Int8(n) => Arc::new(Int8Array::from(vec![*n])) as ArrayRef,
                     GroupByScalar::Int16(n) => Arc::new(Int16Array::from(vec![*n])),
                     GroupByScalar::Int32(n) => Arc::new(Int32Array::from(vec![*n])),
                     GroupByScalar::Int64(n) => Arc::new(Int64Array::from(vec![*n])),
@@ -837,9 +837,7 @@ fn create_batch_from_map(
     Ok(batch)
 }
 
-fn create_accumulators(
-    aggr_expr: &Vec<Arc<dyn AggregateExpr>>,
-) -> Result<AccumulatorSet> {
+fn create_accumulators(aggr_expr: &Vec<Arc<dyn AggregateExpr>>) -> Result<AccumulatorSet> {
     aggr_expr
         .iter()
         .map(|expr| expr.create_accumulator())
@@ -859,9 +857,7 @@ fn finalize_aggregation(
                 .iter()
                 .map(|accumulator| accumulator.state())
                 .map(|value| {
-                    value.map(|e| {
-                        e.iter().map(|v| v.to_array()).collect::<Vec<ArrayRef>>()
-                    })
+                    value.map(|e| e.iter().map(|v| v.to_array()).collect::<Vec<ArrayRef>>())
                 })
                 .collect::<Result<Vec<_>>>()?;
             Ok(a.iter().flatten().cloned().collect::<Vec<_>>())
@@ -1004,8 +1000,7 @@ mod tests {
 
     /// build the aggregates on the data from some_data() and check the results
     async fn check_aggregates(input: Arc<dyn ExecutionPlan>) -> Result<()> {
-        let groups: Vec<(Arc<dyn PhysicalExpr>, String)> =
-            vec![(col("a"), "a".to_string())];
+        let groups: Vec<(Arc<dyn PhysicalExpr>, String)> = vec![(col("a"), "a".to_string())];
 
         let aggregates: Vec<Arc<dyn AggregateExpr>> = vec![Arc::new(Avg::new(
             col("b"),
@@ -1160,16 +1155,14 @@ mod tests {
 
     #[tokio::test]
     async fn aggregate_source_not_yielding() -> Result<()> {
-        let input: Arc<dyn ExecutionPlan> =
-            Arc::new(TestYieldingExec { yield_first: false });
+        let input: Arc<dyn ExecutionPlan> = Arc::new(TestYieldingExec { yield_first: false });
 
         check_aggregates(input).await
     }
 
     #[tokio::test]
     async fn aggregate_source_with_yielding() -> Result<()> {
-        let input: Arc<dyn ExecutionPlan> =
-            Arc::new(TestYieldingExec { yield_first: true });
+        let input: Arc<dyn ExecutionPlan> = Arc::new(TestYieldingExec { yield_first: true });
 
         check_aggregates(input).await
     }