--- a/rust/datafusion/src/physical_plan/hash_aggregate.rs
+++ b/rust/datafusion/src/physical_plan/hash_aggregate.rs
@@ -31,10 +31,7 @@ use crate::error::{DataFusionError, Result};
 use crate::physical_plan::{Accumulator, AggregateExpr};
 use crate::physical_plan::{Distribution, ExecutionPlan, Partitioning, PhysicalExpr};
 
-use arrow::{datatypes::{DataType, Field, Schema, SchemaRef, TimeUnit}, record_batch::RecordBatch};
-use arrow::{
-    array::BooleanArray,
-};
+use arrow::array::BooleanArray;
 use arrow::{
     array::{Array, UInt32Builder},
     error::{ArrowError, Result as ArrowResult},
@@ -46,6 +43,10 @@ use arrow::{
     },
     compute,
 };
+use arrow::{
+    datatypes::{DataType, Field, Schema, SchemaRef, TimeUnit},
+    record_batch::RecordBatch,
+};
 use hashbrown::HashMap;
 use ordered_float::OrderedFloat;
 use pin_project_lite::pin_project;
@@ -53,7 +54,10 @@ use pin_project_lite::pin_project;
 use arrow::array::{TimestampMicrosecondArray, TimestampNanosecondArray};
 use async_trait::async_trait;
 
-use super::{RecordBatchStream, SendableRecordBatchStream, expressions::Column, group_scalar::GroupByScalar};
+use super::{
+    expressions::Column, group_scalar::GroupByScalar, RecordBatchStream,
+    SendableRecordBatchStream,
+};
 
 /// Hash aggregate modes
 #[derive(Debug, Copy, Clone)]