@Dandandan 2021-01-20
![image](https://user-images.githubusercontent.com/163737/105164296-42515780-5b15-11eb-87f0-a042c4287514.png)

@jorgecarleitao while the code is quite a bit faster already, it seems `Array.slice` doesn't really do a zero-copy as promised in the doc (and accounts for about 10% of the total instructions of the program). Maybe you have an idea what could be going on here?