@jorgecarleitao 2021-01-20
Yes, slicing is suboptimal atm. Also, IMO it should not be the `Array` to implement that method, but each implementation individually. I haven't touch that part yet, though.

That analysis is really cool, though. How did you came up with that????? 😮 Any blog post or readme??? :P