@Dandandan 2021-01-20
Thanks @jorgecarleitao makes sense.
I'm planning a blog post soon again, might put some details there.
In short its `cargo profiler` installed from github https://github.com/svenstaro/cargo-profiler with callgrind and `kcachegrind`.