@Dandandan 2021-01-20
This PR in itself is ready for review now. The change in approach will help us in the future, even more with `Array.slice` being improved (will create a ticket later), execution times should be further reduced by something like 15-20%.