@Dandandan 2021-01-22
@jorgecarleitao  I found the "offending" code is this function in `array/data.rs` which does a `self.clone()`. Any idea how we could write a non copying version?
I think it was copied/adapted based on a function in the array module, but I think it was also cloning in previous versions.

```rust
    /// Creates a zero-copy slice of itself. This creates a new [ArrayData]
    /// with a different offset, len and a shifted null bitmap.
    ///
    /// # Panics
    ///
    /// Panics if `offset + length > self.len()`.
    pub fn slice(&self, offset: usize, length: usize) -> ArrayData {
        assert!((offset + length) <= self.len());

        let mut new_data = self.clone();

        new_data.len = length;
        new_data.offset = offset + self.offset;

        new_data.null_count =
            count_nulls(new_data.null_buffer(), new_data.offset, new_data.len);

        new_data
    }
    
 ```