@jorgecarleitao 2021-01-22
Isn't the data contained on a buffer `Arc`ed? I.e. `Vec<Buffer>::clone()` should be cheap, no?