@Dandandan 2021-01-22
@jorgecarleitao 
Maybe, reasonably, yes (as long the underlying vecs have a few items).

I think the clone on the  `ArrayData` structure itself is expensive when slicing for a few elements as it does some allocations for the struct, the vecs, execute some code, etc.