@nevi-me 2021-01-22
> I found the "offending" code is this function in `array/data.rs` which does a `self.clone()`.

This relates to the other discussion that we had on how slicing an array does a clone without propagating offset information to child_data and buffers.

I find it interesting though that `self.clone()` would cost us so much. If you were to clone an array, and inspect the original and the cloned, the buffers point to the same location in memory. I thought that's what zero-copying would give us.

Is that not the case?