@Dandandan 2021-01-22
This PR itself is ready for review.

I think the performance for slicing for small slices would be something to look at later.